move integration tests out of reporter
