+++ src/reporter/version.py
@@ -2,5 +2,5 @@
 
 def version():
     return {
-        'version': '0.8.0'
+        'version': '0.8.x'
     }
