+++ src/tests/run_tests.sh
@@ -11,18 +11,18 @@ tot=0
 echo "\n"
 echo "Launch services with previous CRATE and QL version"
 CRATE_VERSION=${PREV_CRATE} QL_IMAGE=${QL_PREV_IMAGE} docker-compose up -d
-sleep 10
+sleep 20
 
 
-ORION_HOST=`docker inspect -f '{{range .NetworkSettings.Networks}}{{.IPAddress}}{{end}}' $(docker ps | grep "1026" | awk '{ print $1 }')`
-QUANTUMLEAP_HOST=`docker inspect -f '{{range .NetworkSettings.Networks}}{{.IPAddress}}{{end}}' $(docker ps | grep "8668" | awk '{ print $1 }')`
+ORION_BC_HOST=`docker inspect -f '{{range .NetworkSettings.Networks}}{{.IPAddress}}{{end}}' $(docker ps | grep "1026" | awk '{ print $1 }')`
+QL_BC_HOST=`docker inspect -f '{{range .NetworkSettings.Networks}}{{.IPAddress}}{{end}}' $(docker ps | grep "8668" | awk '{ print $1 }')`
 
 # Load data
 echo "\n"
 echo "Load data"
 docker run -ti --rm --network tests_default \
-           -e ORION_URL="http://$ORION_HOST:1026" \
-           -e QL_URL="http://$QUANTUMLEAP_HOST:8668" \
+           -e ORION_URL="http://$ORION_BC_HOST:1026" \
+           -e QL_URL="http://$QL_BC_HOST:8668" \
            --entrypoint "" \
            -e USE_FLASK=TRUE \
            smartsdk/quantumleap python tests/common.py
@@ -36,13 +36,13 @@ sleep 40
 echo "\n"
 echo "Backwards Compatibility Test"
 cd ../../
-pytest src/tests/test_bc.py --cov-report= --cov-config=.coveragerc --cov-append --cov=src/ 
+pytest src/tests/test_bc.py --cov-report= --cov-config=.coveragerc --cov-append --cov=src/
 tot=$?
 
 # Integration Test
 echo "\n"
 echo "Integration Test"
-pytest src/tests/test_integration.py --cov-report= --cov-config=.coveragerc --cov-append --cov=src/ 
+pytest src/tests/test_integration.py --cov-report= --cov-config=.coveragerc --cov-append --cov=src/
 loc=$?
 if [ "$tot" -eq 0 ]; then
   tot=$loc
